First pass at renaming all mentions of annotationsx to hxat
