--- annotationsx/settings/aws.py
@@ -1,3 +1,0 @@
-from .base import *
-
-STATICFILES_STORAGE = "annotationsx.staticfiles.MixedManifestStaticFilesStorage"
