+++ hxat/asgi.py
@@ -9,7 +9,7 @@ from channels.routing import get_default_application
 from dotenv import load_dotenv
 
 # default settings_module is prod
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "annotationsx.settings.prod")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "hxat.settings.prod")
 
 # if dotenv file, load it
 dotenv_path = os.environ.get("HXAT_DOTENV_PATH", None)
