+++ hxat/exceptions.py
