+++ hxat/settings/aws.py
@@ -1,0 +1,3 @@
+from .base import *
+
+STATICFILES_STORAGE = "hxat.staticfiles.MixedManifestStaticFilesStorage"
