+++ hxat/settings/prod.py
@@ -12,4 +12,4 @@ LOGGING["loggers"]["hx_lti_initializer"]["handlers"] = ["console"]
 LOGGING["loggers"]["hx_lti_assignment"]["handlers"] = ["console"]
 LOGGING["loggers"]["target_object_database"]["handlers"] = ["console"]
 LOGGING["loggers"]["annotation_store"]["handlers"] = ["console"]
-LOGGING["loggers"]["annotationsx.middleware"]["handlers"] = ["console"]
+LOGGING["loggers"]["hxat.middleware"]["handlers"] = ["console"]
