+++ locust/requirements.txt
@@ -1,4 +1,4 @@
--r ../annotationsx/requirements/test.txt
+-r ../hxat/requirements/test.txt
 
 locustio==0.14.6
 websocket-client
