+++ requirements.txt
@@ -1,1 +1,1 @@
--r annotationsx/requirements/base.txt
+-r hxat/requirements/base.txt
